fix: check if frame address is within stack chunk
