+++ echion/stacks.h
@@ -20,80 +20,9 @@
 #include <echion/config.h>
 #include <echion/frame.h>
 #include <echion/mojo.h>
+#include <echion/stack_chunk.h>
 
-#if PY_VERSION_HEX >= 0x030b0000
-// ----------------------------------------------------------------------------
-
-class StackChunkError : public std::exception
-{
-public:
-    const char *what() const noexcept override
-    {
-        return "Cannot create stack chunk object";
-    }
-};
-
-// ----------------------------------------------------------------------------
-class StackChunk
-{
-public:
-    StackChunk(PyThreadState *tstate) : StackChunk((_PyStackChunk *)tstate->datastack_chunk) {}
-
-    void* resolve(void *frame_addr);
-
-private:
-    void* origin = NULL;
-    std::unique_ptr<char[]> data = nullptr;
-    std::unique_ptr<StackChunk> previous = nullptr;
-
-    StackChunk(_PyStackChunk *chunk_addr);
-};
-
-// ----------------------------------------------------------------------------
-StackChunk::StackChunk(_PyStackChunk *chunk_addr)
-{
-    _PyStackChunk chunk;
-
-    // Copy the chunk header first
-    if (copy_type(chunk_addr, chunk))
-        throw StackChunkError();
-
-    origin = chunk_addr;
-    data = std::make_unique<char[]>(chunk.size);
-
-    // Copy the full chunk
-    if (copy_generic(chunk_addr, data.get(), chunk.size))
-        throw StackChunkError();
-
-    if (chunk.previous != NULL) {
-        try {
-            previous = std::unique_ptr<StackChunk>{ new StackChunk((_PyStackChunk*)chunk.previous) };
-        }
-        catch (StackChunkError &e) {
-            previous = nullptr;
-        }
-    }
-}
-
-// ----------------------------------------------------------------------------
-void* StackChunk::resolve(void *address)
-{
-    _PyStackChunk *chunk = (_PyStackChunk *)data.get();
-
-    // Check if this chunk contains the address
-    if (address >= origin && address < (char *)origin + chunk->size)
-        return (char *)chunk + ((char *)address - (char *)origin);
-
-    if (previous)
-        return previous->resolve(address);
-
-    return address;
-}
-
-// ----------------------------------------------------------------------------
-
-static std::unique_ptr<StackChunk> stack_chunk = nullptr;
-#endif
+#define MAX_FRAMES 2048
 
 // ----------------------------------------------------------------------------
 
@@ -190,8 +119,8 @@ unwind_frame(PyObject *frame_addr, FrameStack &stack)
 #if PY_VERSION_HEX >= 0x030b0000
             auto resolved_address = stack_chunk ? stack_chunk->resolve(current_frame_addr) : current_frame_addr;
             Frame &frame = resolved_address != current_frame_addr
-                ? Frame::read_local((_PyInterpreterFrame *)resolved_address, &current_frame_addr)
-                : Frame::read((PyObject *)current_frame_addr, &current_frame_addr);
+                               ? Frame::read_local((_PyInterpreterFrame *)resolved_address, &current_frame_addr)
+                               : Frame::read((PyObject *)current_frame_addr, &current_frame_addr);
 #else
             Frame &frame = Frame::read(current_frame_addr, &current_frame_addr);
 #endif
@@ -222,14 +151,17 @@ unwind_frame_unsafe(PyObject *frame, FrameStack &stack)
 
 #if PY_VERSION_HEX >= 0x030d0000
         // See the comment in unwind_frame()
-        while (current_frame != NULL) {
-            if (((_PyInterpreterFrame*)current_frame)->f_executable->ob_type == &PyCode_Type) {
+        while (current_frame != NULL)
+        {
+            if (((_PyInterpreterFrame *)current_frame)->f_executable->ob_type == &PyCode_Type)
+            {
                 break;
             }
             current_frame = (PyObject *)((_PyInterpreterFrame *)current_frame)->previous;
         }
 
-        if (current_frame == NULL) {
+        if (current_frame == NULL)
+        {
             break;
         }
 #endif // PY_VERSION_HEX >= 0x030d0000
@@ -255,16 +187,18 @@ unwind_python_stack(PyThreadState *tstate, FrameStack &stack)
 {
     stack.clear();
 #if PY_VERSION_HEX >= 0x030b0000
-    try {
+    try
+    {
         stack_chunk = std::make_unique<StackChunk>(tstate);
     }
-    catch (StackChunkError &e) {
+    catch (StackChunkError &e)
+    {
         stack_chunk = nullptr;
     }
 #endif
 
 #if PY_VERSION_HEX >= 0x030d0000
-    PyObject* frame_addr = (PyObject*)tstate->current_frame;
+    PyObject *frame_addr = (PyObject *)tstate->current_frame;
 #elif PY_VERSION_HEX >= 0x030b0000
     _PyCFrame cframe;
     _PyCFrame *cframe_addr = tstate->cframe;
@@ -285,16 +219,18 @@ unwind_python_stack_unsafe(PyThreadState *tstate, FrameStack &stack)
 {
     stack.clear();
 #if PY_VERSION_HEX >= 0x030b0000
-    try {
+    try
+    {
         stack_chunk = std::make_unique<StackChunk>(tstate);
     }
-    catch (StackChunkError &e) {
+    catch (StackChunkError &e)
+    {
         stack_chunk = nullptr;
     }
 #endif
 
 #if PY_VERSION_HEX >= 0x030d0000
-    PyObject* frame_addr = (PyObject*)tstate->current_frame;
+    PyObject *frame_addr = (PyObject *)tstate->current_frame;
 #elif PY_VERSION_HEX >= 0x030b0000
     PyObject *frame_addr = (PyObject *)tstate->cframe->current_frame;
 #else // Python < 3.11
